requirements.txt
numpy~=1.23.3
wandb~=0.13.4
tqdm~=4.64.1
scikit-learn~=1.1.2
pandas~=1.5.0
torchaudio~=0.12.1
matplotlib~=3.6.0
scipy~=1.9.1
joblib~=1.2.0